v4.0.272
11-Sep-2012, (New in OTL 4.0.272):

A problem was reported that when an OTL based program has a long uptime (months on end), and the program open and closes OTL streams a lot, dynamic heap fragmentation goes up. In order to address the problem, this release introduces the following #define: OTL_INITIAL_VAR_LIST_SIZE.

This release compiles clean with CLANG++ 3.1.

02-Sep-2012, (New in OTL 4.0.271):

An "unused variable" compilation warning was reported when OTL complies with CLANG. The problem is fixed in this release.

11-Aug-2012, (New in OTL 4.0.270):

A problem was reported when #define OTL_UNICODE is used with Oracle and LONG table columns. The problem is that sometimes LONG string values contain garbage characters at the end, presumably from longer LONG values fetched by the otl_stream before. The problem is fixed in this release.

04-Aug-2012, (New in OTL 4.0.269):

A bug was reported that the OTL header file didn't compile with ACE 6.0 when #define OTL_ACE was enabled. The problem is fixed in this release.

20-Jul-2012, (New in OTL 4.0.268):

This release compiles clean with GNU C++ 4.7.1, which was released over a month ago.